New profile design incl projexp viewer & edit mode
Continued working on minimalist design. We've now got projects in what's nearly a grid on the right side of the page, and a photo, space for personal info, and a nameplate.

== Changes to the profile page ==

The profile page now exhibits a half-arsed edit mode. Visit
</people/username/edit> to see it in bloom.

The profile page now doesn't display a throbber that says "Importing
your commits from 'round the web." This UI change reflects the fact,
made true in an earlier revision, that the profile page controller
doesn't automatically import your commits.

How are commits imported, you ask? Well, there's a bloomin' **form** on this page that cries, "Import commits from around the web". Should you punch in your username or email, and smack the 'add' button, the form does what it promises.

There's also a demi-buttocksed signup form on the profile page.

Tag editing occurs in-place, on the profile page. I can't quite believe
that the tests of tag editing still work, since I don't remember
touching them, but low and behold like an watchful cow, the tests all
say OK.

=== Avatars ===

The UI for the profile page's edit mdoe currently suggests that you can
upload a photo or "doodle". Unfortunately the behind-the-scenes logic is
oblivious to this promise.

== Template reörganization ==

Divided profile into two templates:

1. header--does nothing at the moment.
2. personal--displays personal data

These templates are called from main.html (which displays a grid of ProjectExps) and projectexp.html (which displays details for a single projectexp).

== A new way to display projects ==

When you arrive on the profile page, there's a grid of project icons.
Should you click on one of these icons, you'll be delivered to a page
that talks in details about the profiled user's connection to the
project. Let's call this page the 'ProjectExp view'.

That page allows the logged-in user to delete this information.

It also has a back button which takes the user back to the grid of
project icons.

The ProjectExp view also displays specific details about the profiled user's
involvement with the project, namely:
    - the number of months over which the profiled user contributed commits to the project
    - the primary language of the project.

The view also displays a star icon which when clicked may or may not list the project as one of the profiled user's favorites. I don't remember whether this
works, or whether we disabled the tests.

== Miscellany ==

* Added lots of images.
* Fixed a spurious slash on the signup link.
* Much code folding support for vim. In this respect vim is inferior to NetBeans.
* Add link to the section of the profile that displays in detail information about the profiled user's involvement with a single project. The link takes one back to the section of the profile that displays a list of projects that the profiled user has been involved with.
* Made some cosmetic changes to the code in the testing suite. Rewrote some documentation strings so that they appear on one line. If the doc strings are written like this, they will printed in full to the screen during testing.
* You know the method in the test suite that tests whether paulproteus can log in and remove the word 'ccHost' from his profile by clicking the delete button? I made that method a little easier for me to read. I found some lines of code that were very productive, each accomplishing three or four things by itself. I rewrote each these lines as several lines of code, each of which accomplishes fewer things.
* In the testing method that tests if paulproteus can log in and remove
the word 'ccHost' from his profile by clicking the delete button...
Make twill visit the ProjectExp page instead of just the main profile
page, because the ProjectExp page is now where the delete button is.
Make the code slightly easier to read (in accordance with the
aforementioned principle of write more lines, and do fewer things on
each line).
* Add a testing method that tests that if paulproteus, while logged in, tries to POST directly to the ProjectExp deletion script, without specifying the id of a ProjectExp, the server returns a 500 error message.
* Change the testing method that tests that the ProjectExp page displays the description. The text tested for is now correct.
* View that displays a ProjectExp now passes the appropriate ProjectExp object to the template.
* View that deletes ProjectExp objects will return a 500 error if no id was sent to POST.