Authorization & removing usernames from post URLs
== Authorization ==
We realized that most of the secured parts of the site only concern
operations on the logged-in user.

So authorization is mostly just authentication.

Reminder: authentication means handling login/logout; authorization means limiting
access.

There are two exceptions. We edit ProjectExps and LinkPersonTags.

For ProjectExps, you can't delete one unless it's yours, and there's a
test for that.

For LinkPersonTags, you can't delete one unless it's yours, and there's
no test for that. For the moment, we trust the functionality, but we
know we should go back and fix this at a later stage.

== Cleaning up URLs a bit ==

If it's a URL we POST to, we removed the ?<username> regex expressions,
corresponding parameters in views, and references to those URLs in
templates.

We didn't make exhaustive corrections to Javascript. (We don't yet test Javascript. Sorry,
ziegs!)

== Removing some code ==

Scratched tests and views that we don' use.

== Commit msgs ==

Give the Barry user fixture a password we actually know

The test that Barry cannot delete an exp by someone else now passes

+fixme

New test that we're gonna delete anyway. It tests that you can't edit someone else's tags.

Removed username as a parameter to tags_edit_do.

Removed username from test_commit_importer_json URL; aced CeleryTests.

CommitImportTests (and its view, display_test_page_for_commit_importer) updated to have no username argument.

Removed usernames from URLs concerning the logged-in user.

Use `edit_mode` not `edit` in the data passed to profile templates.
Removed TrentonTests, which concern favorites for ProjectExps and
LinkPersonTags.
What was
    display_person_web
        display_person
            data_for_person...
    display_person_edit_web
        display_person
            data_for_person...

    is now

    display_person_web
        data_for_person...
    display_person_edit_web
        data_for_person...

In the data passed to the profile templates, data['the_user'] = request.user

Tabs are no more.

Removed dead method: display_person_old