Clean ups related to 404 page.
The 404 page knows when a user is logged in, and displays the
user bar at the top of the page.

Simplify content of 404 page.

Use get_or_create in display_person_web. This will create a profile
for a user on profile load if there isn't one already. This makes
setting up a new dev system easier, because then new devs can log
into OpenHatch with their Django superuser account.