Satisfy some requirements of PortfolioEntry.Save's integration test.
The test won't work unless we are trying to save a portfolio entry
that actually exists. This means we should run this test with real
data. For the moment I'll just comment out the test. That's because I
think it's safe to assume that if the Django test and the JS unit
test are both satisfied, then the communication between their
respective logicks should work. Of course, this leaves some
crucial questions unaddressed--e.g., is javascript using the
correct URL for the POST? that's not tested (although it's
currently true).