PeopleMap: Fixed mysterious getBounds problem.
We're having a strange problem: now that we instruct
the items in the list of people to recenter the map
in response to being clicked, the map later complains
later (when we're asking our markers if they're
visible on the map) that its bounds are not set.
Specifically, map.getBounds() returns an undefined
value, where map is a Google Map object (specifically,
one returned by google.maps.Map). To solve this
problem, we center the map at geographic coordinates
0,0 repeatedly during the map initialization process
-- whenever we add a marker. That seems to resolve the
getBounds problem.