Explain the uninhabited island when its marker is clicked.
Don't print multiple messages for multiple clicks, unless the user
triggers the message via a different method.

(The idea of a button displaying a message at least once is easy to
grasp. But I surmise it's not improbable for the user to click the
"Inaccessible Island" marker, then the question mark, in
somewhat quick succession, or vice versa. In this sort of case, we
wouldn't want the user to think that clicking on the second item has
done nothing.)