WIP: Print questions on project pages.
This is failing for the following reason. Check out this bit of template
code from <project/templates/project/project.html>:

    {% for question in project.questions.all %}

Django is parsing the expression "project.questions.all" as a
ManyToManyManager, not a QuerySet. So it throws the exception:

    'ManyRelatedManager' object is not iterable

Frankly, this is NUTSO.

Whereas the following Python code appears to work just fine in the
interactive shell:

    from django.template import Template, Context
    t = Template(
            """
        {% for question in project.questions.all %}
            {{ question.text }}
        {% endfor %}
        """)

    p = Project.objects.get(name='Ubuntu')
    # Ubuntu has a question associated with it.

    print t.render(Context({'project': p}))

    # The output is the value of the Ubuntu question's `text' field:
    # "Why"

What. Teh. Dilly.

Yr. befuddled correspdt.,

Raffi