Move the icon rescaling to the end of the migration process.
The rescaling code depends on later migrations, but the later migrations
don't depend on the rescaling having been done.  Leaving the rescaling
where it is breaks migration when it's all done in one shot (i.e. when
setting up a fresh database).

(Note: The old location is left in but made into a no-op so people
with established databases don't suddenly find south refusing to deal
with a migrations dir that is missing a migration that south knows it
already did.)