Fixed Mercurial bug importer (and hopefully all Roundup bug trackers)
Until this patch series, when a RoundupBugTracker subclass woke up and there was a days-old bug that
it needed to refresh, it would not properly refresh the bug. It would create a new Bug object with
the same canonical_bug_link value as the old bug. When it came to be time to save that Bug object,
the database would balk with an IntegrityError.

Now, if such a Bug exists, we re-use the existing object, overwriting its data.