Gracefully handle when profile photo processing fails.
issue166 describes a semi-corrupt photo that passes Django's image
validation tests, but would cause an uncaught exception during form
validation that resulted in serving a 500. Now, catch the exception
and report a helpful message to the user.

Also add unit tests for a photo that fails the Django image validation
tests and for a photo that causes an error during decompression like
in issue166.