Preserve the line breaks in multi-paragraph project Chatter answers.
Also add a unit test for this behavior, and adjust test_create_answer,
which had carriage returns in the answer text that result in a
different expected display text after this change.

Note that the newlines were preserved in the database and it was the
display code that was not inserting line breaks, so this change
affects how existing multi-paragraph answers are displayed.

Closes: https://openhatch.org/bugs/issue131