openhatch-0.10.11 - Our first release.
Four people contributed code to this release:

* Jessica McKellar
* Asheesh Laroia
* Jack Grigg
* Karen Rustad

It's named 0.10.11 by Jack:

    0: We're too timid to commit to this release naming strategy,
       so if we keep these releases numbered at 0.x, we have lots
       of room to grow into.

    10: The year is 2010.

    11: We started it in month 11, November.

The "started in" thing is confusing, but that's the way we did it.
Going forward we'll start shortening the releases until they start
and end in the same month to avoid this confusion.

(The reason "started in" came about is that we initially thought of
releases as a set of bugs to work on. So it was named for the time we
started putting bugs into that list.)

You might be able to find more about this release on the OpenHatch blog,
although at the time of this writing this is all you get.

(We said we'd ship this on December 13 or 14, but here I am typing it on December
15. The last commit is from December 12.)

-- Asheesh Laroia, 2010-12-15.