Add support for old versions (i.e. 10.*) of Trac.
Usually the date_reported and last_touched fields of a bug are pulled from the
bug page, but in older versions of trac the "Opened' and "Last Modified"
strings are not linked to the timeline. To get around this, we now grab and
store the timeline ticket info for these problem trackers via an RSS feed.
This gives both dates accurately for closed bugs, and date_reported for open
bugs. To get last_touched for open bugs, we then grab the RSS feed of each one
and take the date of the latest comment.