Add Project field for the homepage and add it to info and edit pages.
Side note:
The edit in Bug.title was because for some reason the title field in the
database had max_length=500 but the model itself had max_length=200, and
when I created a migration it shortened the field, throwing a warning of
titles being truncated. So the model has been edited to match the field.

Closes https://openhatch.org/bugs/issue61