Change weekly e-mails to be the maintainer update e-mails.
This is the first step in a substantial overhaul of "weekly" (see
https://openhatch.org/bugs/issue284 for the possibility of daily)
e-mails. Many of the tests will eventually need to be revised or
renamed.

We'll eventually want distinct content for maintainers and
contributors. For now, just filter the projects considered for weekly
e-mails based on maintainer status, fix the tests to use that renamed
function, and remove
test_that_contributors_and_wanna_helpers_are_emailed_about_one_another,
because that's no longer true.