I've added the management command -w to check for the mailing. I'm not sure if the actual database datetime function is ok, but it should at least be a small stepping stone for it's integration.
There have been changes to tests.py names so that all the nagios tests can be
under one class. I've added the weekly email ones so that there is some sort
of proof that the thing works.

Also added the new test methods for the send weekly mail test for a new
database and for a database for which the send weekly mail management command
has been run.

Anyway, this is a little much and I'm seriously tired and slightly drunk, so
I'll leave the hard checking to wiser and more sober individuals.

Cheers.