Change all (I hope) project.name displays to project.display_name
One tricky thing now is that, because of the way /search/ is implemented, the
code in search/views.py and search/controllers.py uses project.display_name
to identify projects, but this is not currently unique. I don't know if this
will cause issues or not (I am not too familiar with the /search/ code) but if
so then the simplest fix is to make project.display_name unique. I don't know
how editing of project.display_name will then go.