Use a patch that can't be applied using -p0 for the recursive patch mission.
The old patch could be applied from the parent directory of the hats
and hats.patch directory with:

patch -p0 < hats.patch

Instead, add leading a/ and b/ directories to hats.patch, imitating
the format of git patches.

This commit also updates the mission instructions and test case to
reflect the new patch.