Git mission: accept password with dot (as shown)
The last git mission tells you to find the password, which is given to you like
this:

    Knock 3 times and give the secret password: Pinky.

But the mission page won't accept "Pinky." as the password, only "Pinky".

Programmer types like to be all literal. I say we encourage that sort of thing.

Noticing punctuation is a good thing!

This patch makes the mission page accept the password with the period at the end.