Fix XSS in map rendering page
If you had a person whose name was a carefully crafted HTML tag, then
the link to the person could include the HTML. This was due to the reckless
use of string concatenation to build HTML from within JavaScript, which is
always wrong.

This fix uses jQuery to build the elements.