Detect Windows style new lines instead of just Unix style new lines and throw an exception if stripped white spaces have been identified.
When a copy paste is done from a Windows terminal, white spaces were being stripped from the generated diff file and this caused the Taining Mission to fail the patch when pasted by the user.

This is fixed by adding an exception to identify stripped white spaces by detecting windows specific new line characters. The fix was also tested with a successful test case for a windows generated diff file input.