Removing Vagrant setup
We used to support using Vagrant to build a Linux-based virtual machine
that could be used as a development environment for oh-mainline, but
we found this was too complex. Instead, we fixed the app so that all
the non-pure-Python dependencies were optional, and the pure-Python
dependencies are all bundled in vendor/.

As a result, we stopped maintaining the Vagrant setup. So, this commit
removes them, including the puppet manifests that support Vagrant.