Add a sponsors page
This moves the language about the prospectus onto the sponsors page,
and it adds a footer link to the sponsors page, and it adds a link to
the sponsors page from the /donate/ page. It also names our sponsors
to date, to whom we are grateful!

It also adds an in-kind sponsors section, where we can thank Linode.