Strip tags before displaying project name
A project name such as '""><script>alert("asd")</script>' was causing
problems as described in http://openhatch.org/bugs/issue875

Use the Django 'striptags' filter to remove the tag at the end (standard
escaping does not seem to catch this for some reason, even after
specifying the 'escape' filter.)

Resulting fix still leaves a mangled project name, like '"">' or such,
but otherwise leaves the list of projects alone. I'm not totally
satisfied with this fix, and think we should have a similar one in the
submission form so project names like this don't get into the database
to begin with.

Signed-off-by: Mandar Gokhale <mandar.mmg@gmail.com>