Skip time-related test if we are running via MySQL
Closes #1341

The thing about this commit is that MySQL itself supports microsecond precision
just fine. But the Django ORM does not support that through MySQL.

Luckily, we still run this test when we operate with sqlite, so it is reasonably
reasonable to skip it in the case that we are in MySQL mode.