Simplify the deploy script
Back in the day, the deploy script would also do a git push. This removes that
from the deploy script, and assumes that whatever is on the main git repository
as master is what you want to deploy.

This also means we can simplify the docs around the deploy script, because after
this change, it no longer depends on the local git configuration.