Use top-level manage.py as authoritative manage.py
This avoids a problem where we play path games and put
bytestrings on sys.path, resulting in pkg_resources.resource_stream
causing a crash in zipimport, which prevents pytz from being loaded.

Plus, it is just simpler.

Before and after this commit, the top-level directory (rather than
mysite/) was treated as the current working directory, so this should
not cause any breakage related to the current working directory.